+++ CHANGES.rst
@@ -25,10 +25,12 @@ issues if those validators, or default values, were misconfigured in the first p
   
 - Add collective.z3cform.norobots integration #145
   [1letter/gomez]
-  
+
 - For CSV and XML attachments send an empty string if the value is None instead of the string "None"
   [nngu6036]
-  
+
+- Put in tests to show recaptcha validation prevents submissions
+  [djay]
 
 2.1.0 (2019-04-25)
 ------------------